Fix handling of sandbox completion checks
User:Toti&bubu was unable to mark a sandbox exercise as complete, because TrainingModulesUsers#eligibile_for_completion was looking for the url-encoded version of the expected sandbox page and failing to find it.

This is a quick fix that works for the affected user, but there might be other username edge cases where this will fail. It would be useful to add tests for various other characters as part of usernames.