Adds checks to prevent errors:
- Prevents `NotFoundError: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.` and `NotFoundError: The object can not be found here.` in training_slide_handler.jsx by wrapping <FastTrainingAlert/> in a div element. The error happens when google translate interferes with the DOM and then the DOM element's native removeChild method crashes.

- Prevents `NotFoundError
Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.` in tag_list.jsx by wrapping the string I18n.t('courses.none') in a <span> element.

- Prevents `TypeError
Cannot read properties of undefined (reading 'setUploadFilters')` by removing this as it is undefined in functional components.

- Prevents `TypeError
Cannot read properties of undefined (reading 'wp10')`, `TypeError: d[0] is undefined` and `TypeError: undefined is not an object (evaluating 'd[0].wp10')` by adding checks to ensure `articleData` is a non-empty array

- Prevents `TypeError: Cannot read properties of undefined (reading 'coordinates')` and `TypeError: Cannot read properties of undefined (reading 'filteredParagraphs')` by adding optional chaining and fallback values when `this.scrollObject[name]` is undefined