Fix JS error with ArticleViewer for assigned article
The revision range query was throwing errors related to looking for the matching article in the articles store, in cases wherew the article is not in the article store (because it was assigned but not edited, so there is no revision range). This just adds some guard statements to avoid that error noise.