Add a check for currently-running updates in ScheduleCourseUpdates
The switch to the :until_and_while_executing sidekiq-unique-jobs option doesn't seem to have worked to prevent adding duplicates of already-running jobs to the queue. I think I had tried this in the past too, and it didn't work. I'm not sure whether it's a bug with the plugin or whether I misunderstand how that lock strategy is supposed to work, but hopefully this will work around the problem by checking for currently-processing jobs before enqueuing another for the same course.