Speed up training tool spec, try to fix flaky test
The `page.has_selector?` needs wait: 0, or else it will take the full timeout for every slide that doesn't trigger the alert.

Most of the `sleep` calls can be avoided by adding expectations that guarantee the a page has loaded or by navigating to a new page before checking database state.

The 'redirects to library index page' example was failing intermittently. I'm not exactly sure why, but hopefully using a browser-driven logout flow will fix it.