Fix overflow bug with MyArticles Progress Tracker UI
The 5-step version could overflow out of the MyArticles box in some situations. In Firefox and Chrome I could only replicate this by making the screen narrow, but I observed in from a screensharing user even with a window wide enough to include margins.

This quick fix will add a scroll within the Progress Tracker component if it's going to overflow.