Filter out tables before checking Pangram
This gets rid of the template instructions, which are are inside of `table` elements. It will also remove other tables, like infoboxes, but since those often contain non-prose content, it's probably going to result in cleaner AiEditAlerts overall, even when tables that the editor added are not included in the plain text sent to Pangram.