Fix broken pageview views
I'm guessing that some historical pageview data was revised to filter out some pageviews that were previously being counted as legitimate, and this broke the tests that were written against very specific numbers.

I've adjusted the specific numbers to the current values that are coming out of the pageview API (for the specific pages and date ranges under test), but also switched to using `be_within` so that these tests if there are other minor revisions of pageview stats.