Add 'mentorship program' question to wizard, for new instructors only
The previous implementation of `only_if` panels didn't cover panels that should be included only if the relevant attribute is false.

I decided to handle this on the frontend by looking for `only_if` attributes that start with `!`, rather than adding a separate `only_if_not` option attribute and filtering on that as well. Not entirely happy with how either version would look, but this is good enough.