Rework Pangram checking to isolate diffs
This will help us cut down on false positives caused by already-present AI text in live articles, and also reduce API credit usage since we won't send the entire text of large articles.

This hasn't been tested thoroughly. It worked for examples of a first edit to a sandbox and a large edit to an existing page.