Don't allow cloning of new courses by default
We want to streamline the consistency of timelines for returning instructors by preventing cloning by default. We can manually enable cloning on a case-by-case basis by adding the tag, and also allow cloning of a course before it has been approved as a path for preparing multiple sections of the same course.

We'll also remove all the cloneable tags via script.